requirements.txt
# ─────────────────────────────────────────────
# Legal NLP Dissertation — Full Requirements
# ─────────────────────────────────────────────

# Core ML / NLP
torch>=2.1.0
transformers>=4.38.0
datasets>=2.18.0
accelerate>=0.27.0
sentencepiece>=0.1.99
tokenizers>=0.15.0

# Legal-BERT + LED
# Models loaded from HuggingFace Hub:
#   nlpaueb/legal-bert-base-uncased   <- Legal-BERT
#   allenai/led-base-16384            <- LED (Longformer Encoder-Decoder)

# RAG & Vector Store
faiss-cpu>=1.7.4
chromadb>=0.4.22
sentence-transformers>=2.3.1
langchain>=0.1.14
langchain-community>=0.0.32

# Data Ingestion
requests>=2.31.0
beautifulsoup4>=4.12.0
pdfminer.six>=20221105
PyMuPDF>=1.23.0         # fitz — faster PDF extraction
lxml>=5.1.0
html2text>=2020.1.16

# Text Processing
spacy>=3.7.0
nltk>=3.8.1
regex>=2023.12.25

# Evaluation
rouge-score>=0.1.2
bert-score>=0.3.13
evaluate>=0.4.1
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.12.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0

# Storage / Utilities
SQLAlchemy>=2.0.0
aiofiles>=23.2.1
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.2
httpx>=0.27.0
